requirements.txt
biopython==1.81
pandas==2.1.0
h5py==3.9.0
numpy==1.26.0
jupyterlab
google-generativeai>=0.8.0
python-dotenv>=1.0.0